Klargjer å hekte på virusskanninga, ved å først ta pdf-opplasting ut til eiga service. Kanskje, kanskje ikkje nødvendig, men kjens naturleg å ta ut alt rundt pdf-opplastinga separat
