+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/BrevService.kt
@@ -5,18 +5,13 @@ import no.nav.etterlatte.brev.brevbaker.EtterlatteBrevKode.TOM_MAL_INFORMASJONSB
 import no.nav.etterlatte.brev.db.BrevRepository
 import no.nav.etterlatte.brev.model.Brev
 import no.nav.etterlatte.brev.model.BrevID
-import no.nav.etterlatte.brev.model.BrevInnhold
 import no.nav.etterlatte.brev.model.BrevInnholdVedlegg
 import no.nav.etterlatte.brev.model.BrevProsessType
 import no.nav.etterlatte.brev.model.BrevkodePar
 import no.nav.etterlatte.brev.model.Mottaker
-import no.nav.etterlatte.brev.model.OpprettNyttBrev
 import no.nav.etterlatte.brev.model.Pdf
 import no.nav.etterlatte.brev.model.Slate
 import no.nav.etterlatte.libs.common.feilhaandtering.UgyldigForespoerselException
-import no.nav.etterlatte.libs.common.person.Folkeregisteridentifikator
-import no.nav.etterlatte.libs.common.sak.Sak
-import no.nav.etterlatte.libs.common.tidspunkt.Tidspunkt
 import no.nav.etterlatte.token.BrukerTokenInfo
 import org.slf4j.LoggerFactory
 
@@ -142,31 +137,6 @@ class BrevService(
 
         val result = db.settBrevSlettet(id, bruker)
         logger.info("Brev med id=$id slettet=$result")
-    }
-
-    fun lagrePdf(
-        sakId: Long,
-        fil: ByteArray,
-        innhold: BrevInnhold,
-        sak: Sak,
-    ): Brev {
-        val brev =
-            db.opprettBrev(
-                OpprettNyttBrev(
-                    sakId = sakId,
-                    behandlingId = null,
-                    soekerFnr = sak.ident,
-                    prosessType = BrevProsessType.OPPLASTET_PDF,
-                    mottaker = Mottaker.tom(Folkeregisteridentifikator.of(sak.ident)),
-                    opprettet = Tidspunkt.now(),
-                    innhold = innhold,
-                    innholdVedlegg = null,
-                ),
-            )
-
-        db.lagrePdf(brev.id, Pdf(fil))
-
-        return brev
     }
 
     private fun sjekkOmBrevKanEndres(brevID: BrevID): Brev {
