+++ apps/etterlatte-brev-api/src/test/kotlin/no/nav/etterlatte/brev/BrevRouteTest.kt
@@ -48,6 +48,7 @@ import kotlin.random.Random
 internal class BrevRouteTest {
     private val mockOAuth2Server = MockOAuth2Server()
     private val brevService = mockk<BrevService>()
+    private val pdfService = mockk<PDFService>()
     private val brevdistribuerer = mockk<Brevdistribuerer>()
     private val tilgangssjekker = mockk<Tilgangssjekker>()
 
@@ -240,6 +241,7 @@ internal class BrevRouteTest {
         runServer(mockOAuth2Server, "api") {
             brevRoute(
                 brevService,
+                pdfService,
                 brevdistribuerer,
                 tilgangssjekker,
             )
